feat: внести поправки

+++ service/src/main/java/ru/practicum/compilations/CompilationRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.practicum.compilations.model.Compilation;
 
-public interface CompilationRepository extends JpaRepository<Compilation, Integer> {
+public interface CompilationRepository extends JpaRepository<Compilation, Long> {
     void deleteById(long compId);
     Compilation getById(long compId);
     Page<Compilation> findByPinned(boolean pinned, Pageable page);
